abbreviation_internal {snakecase}	R Documentation

Internal abbreviation marker, marks abbreviations with an underscore behind. Useful if parsing_option 1 is needed, but some abbreviations need parsing_option 2.

Description

Internal abbreviation marker, marks abbreviations with an underscore behind. Useful if parsing_option 1 is needed, but some abbreviations need parsing_option 2.

Usage

abbreviation_internal(string, abbreviations = NULL)

Arguments

string
A string (for example names of a data frame).

abbreviations
character with (uppercase) abbreviations. This marks abbreviations with an underscore behind (in front of the parsing). Useful if parsing_option 1 is needed, but some abbreviations need parsing_option 2.

Value

A character vector.

Author(s)

Malte Grosser, malte.grosser@gmail.com

[Package snakecase version 0.11.1 Index]