relevant {snakecase}	R Documentation

Internal helper for "lower_upper", "upper_lower". This helper returns a logical vector with TRUE for the first and every second string of those which contain an alphabetic character

Description

Internal helper for "lower_upper", "upper_lower". This helper returns a logical vector with TRUE for the first and every second string of those which contain an alphabetic character

Usage

relevant(string)

Arguments

string
A string (for example names of a data frame).

Value

A logical vector.

Author(s)

Malte Grosser, malte.grosser@gmail.com

[Package snakecase version 0.11.1 Index]