replace_special_characters_internal {snakecase}	R Documentation

Internal helper to replace special characters.

Description

Internal helper to replace special characters.

Usage

replace_special_characters_internal(string, transliterations, case)

Arguments

string
A string (for example names of a data frame).

transliterations
A character vector (if not NULL). The entries of this argument need to be elements of stringi::stri_trans_list() (like "Latin-ASCII", which is often useful) or names of lookup tables (currently only "german" is supported). In the order of the entries the letters of the input string will be transliterated via stringi::stri_trans_general() or replaced via the matches of the lookup table. When named character elements are supplied as part of 'transliterations', anything that matches the names is replaced by the corresponding value. You should use this feature with care in case of case = "parsed", case = "internal_parsing" and case = "none", since for upper case letters, which have transliterations/replacements of length 2, the second letter will be transliterated to lowercase, for example Oe, Ae, Ss, which might not always be what is intended. In this case you can make usage of the option to supply named elements and specify the transliterations yourself.

case
Length one character, from the input options of to_any_case.

Value

A character vector.

Author(s)

Malte Grosser, malte.grosser@gmail.com

[Package snakecase version 0.11.1 Index]