General case conversion

Function to convert strings to any case

to_any_case(string, case = c("snake", "small_camel", "big_camel",
  "screaming_snake", "parsed"), preprocess = NULL, protect = NULL,
  replace_special_characters = FALSE, postprocess = NULL, prefix = "",
  postfix = "", unique_sep = NULL, empty_fill = NULL, parsingoption = 1)

Arguments

string
A string (for example names of a data frame).

case
The desired target case, provided as one of "snake", "small_camel", "big_camel", "screaming_snake" or "parsed". The latter one is not really a case, but is helpful since it returns the parsed input string, separated by underscores, without any further modification.

preprocess
String that will be wrapped internally into stringr::regex(). All matches will be treated as additional splitting parameters besides the default ones ("_" and " "), when parsing the input string.

protect
A string which is a valid stringr::regex(). Matches within the output won't have any "_" (or artifacts of preprocess) beside. Note that preprocess has a higher precedence than protect, which means that it doesn't make sense to protect sth. which is already replaced via preprocess.

replace_special_characters
Logical, if TRUE, special characters will be translated to characters which are more likely to be understood by different programs. For example german umlauts will be translated to ae, oe, ue etc.

postprocess
String that will be used as separator. The defaults are "_" and "", regarding the specified case.

prefix
prefix (string).

postfix
postfix (string).

unique_sep
A string. If it is supplied, then duplicated names will get a suffix integer in the order of their appearance. The suffix is separated by the supplied string to this argument.

empty_fill
A string. If it is supplied, then each entry that matches "" will be replaced by the supplied string to this argument.

parsingoption
An integer (1 (default), 2 or 3) that will determine the parsingoption. 1: RRRStudio -> RRR_Studio 2: RRRStudio -> RRRS_tudio If another integer is supplied, no parsing regarding the pattern of upper- and lowercase will appear.

Value

A character vector according the specified parameters above.

Note

to_any_case() is vectorised over postprocess, prefix and postfix. postprocess might follow in the future.

See also

snakecase on github or caseconverter for some handy shortcuts.

Examples

### Default usage
strings <- c("this Is a Strange_string", "AND THIS ANOTHER_One")
to_any_case(strings, case = "snake")

#> [1] "this_is_a_strange_string" "and_this_another_one"    

to_any_case(strings, case = "small_camel")

#> [1] "thisIsAStrangeString" "andThisAnotherOne"   

to_any_case(strings, case = "big_camel")

#> [1] "ThisIsAStrangeString" "AndThisAnotherOne"   

to_any_case(strings, case = "screaming_snake")

#> [1] "THIS_IS_A_STRANGE_STRING" "AND_THIS_ANOTHER_ONE"    

to_any_case(strings, case = "parsed")

#> [1] "this_Is_a_Strange_string" "AND_THIS_ANOTHER_One"    

### Pre -and postprocessing
strings2 <- c("this - Is_-: a Strange_string", "AND THIS ANOTHER_One")
to_snake_case(strings2)

#> [1] "this_-_is_-_:_a_strange_string" "and_this_another_one"          

to_any_case(strings2, case = "snake", preprocess = "-|\\:")

#> [1] "this_is_a_strange_string" "and_this_another_one"    

to_any_case(strings2, case = "snake", preprocess = "-|\\:", postprocess = " ")

#> [1] "this is a strange string" "and this another one"    

to_any_case(strings2, case = "big_camel", preprocess = "-|\\:", postprocess = "//")

#> [1] "This//Is//A//Strange//String" "And//This//Another//One"     

### Pre -and postfix
to_any_case(strings2, case = "big_camel", preprocess = "-|\\:", postprocess = "//",
            prefix = "USER://", postfix = ".exe")

#> [1] "USER://This//Is//A//Strange//String.exe"
#> [2] "USER://And//This//Another//One.exe"     

### Special characters like german umlauts for example can be replaced via 
# replace_special_characters = TRUE

### Protect anything that shouldn't have an underscore beside in the output
strings3 <- c("var12", "var1.2", "va.r.1.2")
to_any_case(strings3, case = "snake")

#> [1] "var_12"         "var_1_._2"      "va_._r_._1_._2"

to_any_case(strings3, case = "snake", protect = "\\d")

#> [1] "var12"       "var1.2"      "va_._r_.1.2"

to_any_case(strings3, case = "snake", protect = "\\d|\\.")

#> [1] "var12"    "var1.2"   "va.r.1.2"