to_parsed_case_internal {snakecase}	R Documentation

Internal parser, which is relevant for preprocessing, parsing and parsing options

Description

Internal parser, which is relevant for preprocessing, parsing and parsing options

Usage

to_parsed_case_internal(string, parsing_option = 1L, numerals,
  abbreviations, sep_in)

Arguments

string
A string.

parsing_option
An integer that will determine the parsing option.

1: RRRStudio -> RRR_Studio

2: RRRStudio -> RRRS_tudio

3: parses like option 1 but suppresses "_" around non alpha-numeric characters. In this way this option suppresses splits and resulting case conversion after these characters.

any other integer <= 0: no parsing"

numerals
A character specifying the alignment of numerals ("middle", left, right or asis). I.e. numerals = "left" ensures that no output separator is in front of a digit.

abbreviations
A character string specifying abbreviations that should be marked to be recognized by later parsing.

sep_in
A character (regular expression) used to specify input separators.

Value

A character vector separated by underscores, containing the parsed string.

Author(s)

Malte Grosser, malte.grosser@gmail.com

[Package snakecase version 0.11.1 Index]