Specific case converter shortcuts

Wrappers around to_any_case()

to_snake_case(string, preprocess = NULL, replace_special_characters = NULL,
  postprocess = NULL, prefix = "", postfix = "", unique_sep = NULL,
  empty_fill = NULL, parsing_option = 1, abbreviations = NULL)

to_lower_camel_case(string, preprocess = NULL,
  replace_special_characters = NULL, postprocess = NULL, prefix = "",
  postfix = "", unique_sep = NULL, empty_fill = NULL,
  parsing_option = 1, abbreviations = NULL)

to_upper_camel_case(string, preprocess = NULL,
  replace_special_characters = NULL, postprocess = NULL, prefix = "",
  postfix = "", unique_sep = NULL, empty_fill = NULL,
  parsing_option = 1, abbreviations = NULL)

to_screaming_snake_case(string, preprocess = NULL,
  replace_special_characters = NULL, postprocess = NULL, prefix = "",
  postfix = "", unique_sep = NULL, empty_fill = NULL,
  parsing_option = 1, abbreviations = NULL)

to_parsed_case(string, preprocess = NULL, replace_special_characters = NULL,
  postprocess = NULL, prefix = "", postfix = "", unique_sep = NULL,
  empty_fill = NULL, parsing_option = 1, abbreviations = NULL)

to_mixed_case(string, preprocess = NULL, replace_special_characters = NULL,
  postprocess = NULL, prefix = "", postfix = "", unique_sep = NULL,
  empty_fill = NULL, parsing_option = 1, abbreviations = NULL)

to_lower_upper_case(string, preprocess = NULL,
  replace_special_characters = NULL, postprocess = NULL, prefix = "",
  postfix = "", unique_sep = NULL, empty_fill = NULL,
  parsing_option = 1, abbreviations = NULL)

to_upper_lower_case(string, preprocess = NULL,
  replace_special_characters = NULL, postprocess = NULL, prefix = "",
  postfix = "", unique_sep = NULL, empty_fill = NULL,
  parsing_option = 1, abbreviations = NULL)

Arguments

string
A string (for example names of a data frame).

preprocess
A string (if not NULL) that will be wrapped internally into stringr::regex(). All matches will be replaced by underscores. Underscores can later turned into another separator via postprocess.

replace_special_characters
A character vector (if not NULL). The entries of this argument need to be elements of stringi::stri_trans_list() (like "Latin-ASCII", which is often useful) or names of lookup tables (currently only "german" is supported). In the order of the entries the letters of the input string will be transliterated via stringi::stri_trans_general() or replaced via the matches of the lookup table.

You should use this feature with care in case of case = "parsed" and case = "none", since for upper case letters, which have transliterations/replacements of length 2, the second letter will be transliterated to lowercase, for example Oe, Ae, Ss, which might not always be what is intended.

postprocess
String that will be used as separator. The defaults are "_" and "", regarding the specified case.

prefix
prefix (string).

postfix
postfix (string).

unique_sep
A string. If not NULL, then duplicated names will get a suffix integer in the order of their appearance. The suffix is separated by the supplied string to this argument.

empty_fill
A string. If it is supplied, then each entry that matches "" will be replaced by the supplied string to this argument.

parsing_option
An integer that will determine the parsing_option. #'

1: RRRStudio -> RRR_Studio

2: RRRStudio -> RRRS_tudio

3: parses at the beginning like option 1 and the rest like option 2.

4: parses at the beginning like option 2 and the rest like option 1.

5: parses like option 1 but suppresses "_" around non special characters. In this way case conversion won't apply after these characters. See examples.

any other integer <= 0: no parsing"

abbreviations
character with (uppercase) abbreviations. This marks abbreviations with an underscore behind (in front of the parsing). useful if parsinoption 1 is needed, but some abbreviations need parsing_option 2.

Value

A character vector according the specified parameters above.

A character vector according the specified target case.

Note

caseconverters are vectorised over postprocess, prefix and postfix. postprocess might follow in the future.

See also

snakecase on github, to_any_case for flexible high level conversion.

Examples

strings <- c("this Is a Strange_string", "AND THIS ANOTHER_One", NA)

to_snake_case(strings)

#> [1] "this_is_a_strange_string" "and_this_another_one"    
#> [3] NA                        

to_small_camel_case(strings)

#> Error in to_small_camel_case(strings): could not find function "to_small_camel_case"

to_big_camel_case(strings)

#> Error in to_big_camel_case(strings): could not find function "to_big_camel_case"

to_screaming_snake_case(strings)

#> [1] "THIS_IS_A_STRANGE_STRING" "AND_THIS_ANOTHER_ONE"    
#> [3] NA                        

to_lower_upper_case(strings)

#> [1] "thisISaSTRANGEstring" "andTHISanotherONE"    NA                    

to_upper_lower_case(strings)

#> [1] "THISisAstrangeSTRING" "ANDthisANOTHERone"    NA                    

to_parsed_case(strings)

#> [1] "this_Is_a_Strange_string" "AND_THIS_ANOTHER_One"    
#> [3] NA                        

to_mixed_case(strings)

#> [1] "this_Is_a_Strange_string" "And_This_Another_One"    
#> [3] NA