Internal helper to test the design rules for any string and setting of to_any_case()

Internal helper to test the design rules for any string and setting of to_any_case()

check_design_rule(string, preprocess = NULL, transliterations = FALSE,
  postprocess = NULL, prefix = "", postfix = "", unique_sep = NULL,
  empty_fill = NULL, parsing_option = 1)

Arguments

string
A string (for example names of a data frame).

preprocess
String that will be wrapped internally into stringr::regex(). All matches will be treated as additional splitting parameters besides the default ones ("_" and " "), when parsing the input string.

transliterations
Logical, if TRUE, special characters will be translated to characters which are more likely to be understood by different programs. For example german umlauts will be translated to ae, oe, ue etc.

postprocess
String that will be used as separator. The defaults are "_" and "", regarding the specified case.

prefix
prefix (string).

postfix
postfix (string).

unique_sep
A string. If it is supplied, then duplicated names will get a suffix integer in the order of their appearance. The suffix is separated by the supplied string to this argument.

empty_fill
A string. If it is supplied, then each entry that matches "" will be replaced by the supplied string to this argument.

parsing_option
An integer (1 (default), 2 or 3) that will determine the parsing option. 1: RRRStudio -> RRR_Studio 2: RRRStudio -> RRRS_tudio If another integer is supplied, no parsing regarding the pattern of upper- and lowercase will appear.

Value

A character vector separated by underscores, containing the parsed string.