Internal parser, which is relevant for preprocessing, parsing and parsingoptions

Internal parser, which is relevant for preprocessing, parsing and parsingoptions

to_parsed_case_internal(string, preprocess = NULL, parsingoption = 1L)

Arguments

string
A string.

preprocess
Character string that will be wrapped internally into stringr::regex. All matches will be padded with underscores.

parsingoption
An integer (1 (default), 2 or 3) that will determine the parsingoption. 1: RRRStudio -> RRR_Studio 2: RRRStudio -> RRRS_tudio if another integer is supplied, no parsing regarding the pattern of upper- and lowercase will appear.

Value

A character vector separated by underscores, containing the parsed string.